\name{Excercise7-package}
\alias{Excercise7-package}
\alias{Excercise7}
\docType{package}
\title{
Set of add functions
}
\description{
Contains 2 functions AddTwo and AddFour that adds 2 or 4 to its arguments
}
\details{
\tabular{ll}{
Package: \tab Excercise7\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-28\cr
License: \tab GPL-2\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Katie Romanenko

Maintainer: Katie Romanenko<katie.romanenko@gmail.com>
}
\references{
http://smaa.fi/static/erimprog/2014/erimprog-2014-ex7.pdf
}
\keyword{ Excercise7 }
\keyword{ AddTwo }
\keyword{ AddFour }
\examples{
a <- 5
b <- -3.4
c <- AddFour(a)+AddTwo(b)
}
